+++ source/OffsetGenerator.h
@@ -179,10 +179,13 @@ class OffsetGenRandom : public OffsetGenerator
 	public:
 		OffsetGenRandom(uint64_t numBytesTotal, RandAlgoInterface& randAlgo, uint64_t len,
 			uint64_t offset, size_t blockSize) :
-			randRange(randAlgo, offset, offset + len - blockSize),
+			randRange(randAlgo, offset, offset + len - std::min(blockSize, len) ),
 			numBytesTotal(numBytesTotal), numBytesLeft(numBytesTotal),
 			blockSize(blockSize)
-		{ }
+		{
+			/* note on "std::min(blockSize, len)": usually blockSize, but there are cases where we
+				have custom tree slices that are smaller than blockSize */
+		}
 
 		virtual ~OffsetGenRandom() {}
 
@@ -200,10 +203,13 @@ class OffsetGenRandom : public OffsetGenerator
 
 		virtual void reset(uint64_t len, uint64_t offset) override
 		{
+			size_t minLenAndBlockSize = std::min(blockSize, len); /* usually blockSize, but there
+				are cases where we have custom tree slices that are smaller than blockSize */
+
 			numBytesTotal = len;
 			numBytesLeft = len;
 
-			randRange.reset(offset, offset + len - blockSize);
+			randRange.reset(offset, offset + len - minLenAndBlockSize);
 		}
 
 		virtual uint64_t getNextOffset() override
@@ -239,11 +245,14 @@ class OffsetGenRandomAligned : public OffsetGenerator
 	public:
 		OffsetGenRandomAligned(uint64_t numBytesTotal, RandAlgoInterface& randAlgo, uint64_t len,
 			uint64_t offset, size_t blockSize) :
-			randRange(randAlgo, 0, (len - blockSize) / blockSize),
+			randRange(randAlgo, 0, (len - std::min(blockSize, len) ) / std::min(blockSize, len) ),
 			numBytesTotal(numBytesTotal), numBytesLeft(numBytesTotal),
 			offset(offset),
 			blockSize(blockSize)
-		{ }
+		{
+			/* note on "std::min(blockSize, len)": usually blockSize, but there are cases where we
+				have custom tree slices that are smaller than blockSize */
+		}
 
 		virtual ~OffsetGenRandomAligned() {}
 
@@ -262,11 +271,14 @@ class OffsetGenRandomAligned : public OffsetGenerator
 
 		virtual void reset(uint64_t len, uint64_t offset) override
 		{
+			size_t minLenAndBlockSize = std::min(blockSize, len); /* usually blockSize, but there
+				are cases where we have custom tree slices that are smaller than blockSize */
+
 			this->numBytesTotal = len;
 			this->numBytesLeft = len;
 			this->offset = offset;
 
-			randRange.reset(0, (len - blockSize) / blockSize);
+			randRange.reset(0, (len - minLenAndBlockSize) / minLenAndBlockSize);
 		}
 
 		virtual uint64_t getNextOffset() override